🧹 코드 건강 개선: email_owner_filters 로직 중복 제거
여러 API(search, emails, ontology) 및 서비스(threading_service)에서
중복 정의되어 있던 email_owner_filters 로직을
단일 모듈(db.filters)로 추출하여 관리하도록 리팩토링했습니다.

- `backend/db/filters.py` 신규 생성
- `email_owner_filters`를 공유 함수로 일원화
- 관련된 모든 파일의 import와 호출부 수정
