+++ backend/api/emails.py
@@ -23,6 +23,7 @@ from services.reply_tracking_service import (
     thread_requires_reply,
 )
 from services.threading_service import normalize_message_id
+from db.filters import email_owner_filters
 from services.email_dedupe_service import (
     EmailDedupeCandidate,
     candidate_message_lookup_values,
@@ -88,14 +89,6 @@ def thread_lookup_values(thread_id: str) -> list[str]:
     normalized = normalize_message_id(thread_id) or thread_id
     return list({thread_id, normalized, f"<{normalized}>"})
 
-
-def email_owner_filters(auth_context: AuthContext):
-    organization_filter = (
-        Email.organization_id == auth_context.organization_id
-        if auth_context.organization_id is not None
-        else Email.organization_id.is_(None)
-    )
-    return (Email.user_id == auth_context.user_id, organization_filter)
 
 
 MailFolder = Literal["inbox", "sent"]
@@ -285,7 +278,7 @@ async def get_emails(
     candidate_window = min(max(limit * 10, 200), 2000)
     result = await db.execute(
         select(Email)
-        .where(*email_owner_filters(auth_context))
+        .where(*email_owner_filters(auth_context.user_id, auth_context.organization_id))
         .order_by(Email.date.desc())
         .limit(candidate_window)
     )
@@ -412,7 +405,7 @@ async def _fetch_existing_emails_for_candidates(
 
     result = await db.execute(
         select(Email).where(
-            *email_owner_filters(auth_context),
+            *email_owner_filters(auth_context.user_id, auth_context.organization_id),
             or_(*predicates),
         )
     )
@@ -610,7 +603,7 @@ async def get_email(
 ):
     # Ensure auth context validates the request payload and scopes access
     result = await db.execute(
-        select(Email).where(Email.id == email_id, *email_owner_filters(auth_context))
+        select(Email).where(Email.id == email_id, *email_owner_filters(auth_context.user_id, auth_context.organization_id))
     )
     email = result.scalar_one_or_none()
     if not email:
@@ -631,7 +624,7 @@ async def get_email_thread(
     result = await db.execute(
         select(Email)
         .where(
-            *email_owner_filters(auth_context),
+            *email_owner_filters(auth_context.user_id, auth_context.organization_id),
             or_(
                 Email.thread_id.in_(lookup_values), Email.message_id.in_(lookup_values)
             ),
